--- a/patrickstar/core/chunk_data.py
+++ b/patrickstar/core/chunk_data.py
@@ -206,7 +206,7 @@ class Chunk(object):
         if self._state_dict[TensorState.COMPUTE] > 0:
             return ChunkState.COMPUTE
         elif self._state_dict[TensorState.HOLD_AND_TOUCHED]:
-            return ChunkState.HOHOLD_AND_TOUCHEDLD
+            return ChunkState.HOLD_AND_TOUCHED
         elif self._state_dict[TensorState.HOLD] > 0:
             return ChunkState.HOLD
         elif self._state_dict[TensorState.HOLD_AFTER_FWD] > 0: