--- a/patrickstar/core/chunk_data.py
+++ b/patrickstar/core/chunk_data.py
@@ -205,7 +205,7 @@ class Chunk(object):
         # Distributed training need to fix the chunk on the compute device.
         if self._state_dict[TensorState.COMPUTE] > 0:
             return ChunkState.COMPUTE
-        elif self._state_dict[TensorState.HOLD_AND_TOUCHED]:
+        elif self._state_dict[TensorState.HOLD_AND_TOUCHED] > 0:
             return ChunkState.HOLD_AND_TOUCHED
         elif self._state_dict[TensorState.HOLD] > 0:
             return ChunkState.HOLD