--- a/patrickstar/core/const.py
+++ b/patrickstar/core/const.py
@@ -48,7 +48,6 @@ class ChunkState(Enum):
 
     # Chunk memory is not allocated.
     RELEASED = 5
-    HOLD_AND_TOUCHED = 6
 
 
 class TensorState(Enum):
@@ -65,8 +64,6 @@ class TensorState(Enum):
     HOLD = 2
     HOLD_AFTER_FWD = 3
     HOLD_AFTER_BWD = 4
-
-    HOLD_AND_TOUCHED = 5
 
 
 class TrainingStage(Enum):