Memory saving communication (MSC): using one-to-all communication to replace the original collective communication. More specifically, reduce scatter is replaced with Nx reduce. all gather is replaced with Nx bcast. In this way, we do not need to keep a Nx chunk buffer for distributed training, therefore save the GPU memory. I have implemented MSC but may not be optimal. I triggered communication in the granularity of the chunk group (Nx chunk). This may lead to chunks frequently move between CPU and GPU.

This MR further optimizes the MSC. Communication is triggered in the granularity of chunks.