@feifeibear 2021-11-26
Case: log.GPT_DS_40B_gpu_8_cs_384_bs_8_cpueb_0_lightseq_0_offload_0_SP_0_AMM_1_MSC_1_CACHE_1

Before Optimization
Step 4 elaspe 96.44639992713928 s, **27.585423709890733** Tflops
CLIENT_fetch_remote_chunks_broadcast .. 0.02838444709777832, 0.02943127151149553 %
CHUNK_LIST_prepare_device ............. 40.753575801849365, 42.256576298871195 %
CHUNK_allocate_payload_cuda ........... 4.351799488067627, 4.512294773814017 %
CLIENT_fetch_remote_chunks ............ 46.959617137908936, 48.69149775229736 %
CLIENT_access_dist .................... 82.38736820220947, 85.4258317706167 %
CLIENT_release_dist ................... 2.8587543964385986, 2.964185863350898 %
chunk_gpu_cpu_move .................... 40.686315298080444, 42.18683521349427 %
CHUNK_LIST_chunk_move ................. 40.69334030151367, 42.19411930054026 %
chunk_cpu_gpu_move .................... 39.72009992599487, 41.18498561408259 %
FWD ................................... 20.306491136550903, 21.05539882552038 %
CLIENT_release_dist_reduce ............ 0.008373260498046875, 0.008682068116584403 %
BWD ................................... 67.8014886379242, 70.30202188234678 %
ADAM_prepare_data_grad_copy ........... 1.113138198852539, 1.1541909711114442 %
CLIENT_access ......................... 0.012293577194213867, 0.01274696692185198 %
ADAM_prepare_data ..................... 1.1289987564086914, 1.1706364693855804 %
ADAM_compute .......................... 4.184556245803833, 4.3388835653951014 %
ADAM_param_fp32_to_fp16 ............... 2.553032398223877, 2.6471887731662984 %
CLIENT_release ........................ 0.015923261642456055, 0.016510514900407818 %
ADAM_release_data ..................... 0.0175015926361084, 0.01814705507500317 %
ADAM .................................. 8.335176229476929, 8.642579292132842 %
TOTAL ................................. 96.44315600395203
------------- DATA MOVE RESULTS --------------
chunk_cpu_gpu_move: 681984.0 MB, 888 times, 17169.74532467565 MB/s
chunk_gpu_cpu_move: 740352.0 MB, 964 times, 18196.585131289325 MB/s
ADAM_prepare_data_grad_copy: 49831.09375 MB, 525 times, 44766.31365392689 MB/s
ADAM_param_fp32_to_fp16: 99662.1875 MB, 525 times, 39036.78918032303 MB/s


After Optimization
Step 4 elaspe 27.47564697265625 s, **96.83174375953567** Tflops
CLIENT_fetch_remote_chunks_broadcast .. 0.023209810256958008, 0.0844838925356625 %
CLIENT_access_dist .................... 9.166956186294556, 33.367792874991835 %
CLIENT_release_dist ................... 0.30394816398620605, 1.1063737160420948 %
CHUNK_LIST_prepare_device ............. 4.056913137435913, 14.767195842409045 %
CHUNK_allocate_payload_cuda ........... 0.0031392574310302734, 0.01142692182782636 %
chunk_gpu_cpu_move .................... 4.043227195739746, 14.717378906608733 %
CHUNK_LIST_chunk_move ................. 4.0436718463897705, 14.71899743848455 %
FWD ................................... 3.798264980316162, 13.82571450892415 %
chunk_cpu_gpu_move .................... 4.971297025680542, 18.095560413060443 %
CLIENT_release_dist_reduce ............ 0.009317398071289062, 0.033915402523844014 %
BWD ................................... 14.895630121231079, 54.220211215893045 %
ADAM_prepare_data_grad_copy ........... 1.465179204940796, 5.333263870978727 %
CLIENT_access ......................... 0.015065431594848633, 0.054838289920142766 %
ADAM_prepare_data ..................... 1.4846744537353516, 5.404226730471363 %
ADAM_compute .......................... 3.7023797035217285, 13.476691344548014 %
ADAM_param_fp32_to_fp16 ............... 3.3249595165252686, 12.10287888481713 %
CLIENT_release ........................ 0.018674373626708984, 0.06797486740231848 %
ADAM_release_data ..................... 0.0205535888671875, 0.07481522571073554 %
ADAM .................................. 8.778572797775269, 31.95407427518281 %
TOTAL ................................. 27.47246789932251
------------- DATA MOVE RESULTS --------------
chunk_cpu_gpu_move: 109056.0 MB, 142 times, 21937.13218836906 MB/s
chunk_gpu_cpu_move: 94464.0 MB, 123 times, 23363.515188939793 MB/s
ADAM_prepare_data_grad_copy: 49831.09375 MB, 525 times, 34010.23818926883 MB/s
ADAM_param_fp32_to_fp16: 99662.1875 MB, 525 times, 29973.955172889277 MB/s