@feifeibear 2021-11-26
result check has been done with.

env RELEASE_AFTER_INIT=1 RES_CHECK=1 GPU_NUM=2 BS=8 CPU_EBD=0 ACT_OFFLOAD=0 AMM=1 MSC=1 CACHE=1 CS=128 bash run_bert.sh

-------------------- LOSS --------------------
torch fp32 : [0.677920401096344, 0.8147394061088562, 10.70045280456543, 5.704251289367676, 7.606137752532959]
autocast   : [0.677734375, 0.8148193359375, 10.703125, 5.705078125, 7.595703125]
patrickstar: [0.6787109375, 0.81396484375, 10.703125, 5.703125, 7.59375]
-------------------- DIFF --------------------
vs torch fp32: [0.0007905364036560059, -0.0007745623588562012, 0.0026721954345703125, -0.0011262893676757812, -0.012387752532958984]
vs autocast  : [0.0009765625, -0.0008544921875, 0.0, -0.001953125, -0.001953125]