@zhuzilin 2021-11-26
revert the async copy for the moment...